Organizing files and folders management, adding README
