+++ Sensors test/DHT/2/Adafruit_DHT.py
