+++ Sensors test/DHT/DHT.pdf
