+++ Sensors test/DHT/DHT11.py
