+++ Sensors test/DHT/setup.py
