+++ Sensors test/Moisture/SPI.py
