+++ Sensors test/SOIL/testSOIL.py
